pyproject.toml
[project]
name = "project-4-debezium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka==2.13.0",
    "duckdb==1.4.3",
    "python-dotenv==1.2.1",
]
